refactor: expose the command layer as a library (lib target + thin bins)
Allow embedding aube's command layer as a library: an embedder can
construct commands::install::InstallOptions and call
commands::install::run in-process instead of shelling out to the CLI.
No behavior change; aube/aubr/aubx still build and dispatch identically.

- move src/main.rs wholesale to src/lib.rs; fn main becomes pub fn
  cli_main, and the new src/main.rs is a thin wrapper that only pins the
  mimalloc global allocator (a binary-level policy the library must not
  impose on embedders) and forwards to aube::cli_main. The aubr/aubx
  multicall shims keep include!-ing src/main.rs unchanged.
- keep the public lib surface small: commands, cli_args, and cli_main;
  every other top-level module stays crate-private.
- visibility promotions forced by rustc's private_interfaces lint once
  the lib target exists: cli_args::{LockfileArgs,NetworkArgs,
  VirtualStoreArgs} and commands::config::{KeyArgs,Location,ListLocation}
  pub(crate) -> pub.
- add examples/embed_smoke.rs, compile-proving the seam.

(reproduces b15cdcb against the current upstream/main main.rs)
